pyproject.toml
[tool.poetry]
name = "triss"
version = "0.1.0"
description = "Triss: an app that looks after your pets using ML."
authors = ["oliyyaa <olga.chaganova1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "4.6.0.66"
onnxruntime = "^1.15.1"
pydantic = "^2.3.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.290"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
select = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

#ignore = ["D100", "UP006", "ANN101", "ANN204", "RUF002", "INP001", "D203", "D213"]

line-length = 120

target-version = "py310"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.flake8-quotes]
inline-quotes = "single"

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "D104"]
"src/routers/face_detection.py" = ["B008"]
"src/services/fdetector/utils.py" = ["TRY003", "EM102", "FBT001", "FBT002"]
"src/repos/cache_service.py" = ["S301"]


[tool.mypy]
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.mypy-numpy]
warn_return_any = false

[tool.mypy-cv2]
warn_return_any = true


[tool.isort]
profile = "hug"
src_paths = ["src"]